Add support for earliest conduit version available
